// @types/schema.d.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** An ISO-8601 encoded date string. */
    Date: any;
    /** An ISO-8601 encoded UTC date string. */
    DateTime: any;
    /** A Git object ID. */
    GitObjectID: any;
    /** Git SSH string */
    GitSSHRemote: any;
    /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
    GitTimestamp: any;
    /** A string containing HTML code. */
    HTML: any;
    /** An ISO-8601 encoded UTC date string with millisecond precision. */
    PreciseDateTime: any;
    /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
    URI: any;
    /** A valid x509 certificate string */
    X509Certificate: any;
};
// ...
export type Actor = {
    /** A URL pointing to the actor's public avatar. */
    avatarUrl: Scalars['URI'];
    /** The username of the actor. */
    login: Scalars['String'];
    /** The HTTP path for this actor. */
    resourcePath: Scalars['URI'];
    /** The HTTP URL for this actor. */
    url: Scalars['URI'];
};
// ...
export type Comment = {
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** The body as Markdown. */
    body: Scalars['String'];
    /** The body rendered to HTML. */
    bodyHTML: Scalars['HTML'];
    /** The body rendered to text. */
    bodyText: Scalars['String'];
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars['DateTime'];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars['Boolean'];
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    id: Scalars['ID'];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars['Boolean'];
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars['DateTime']>;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars['DateTime']>;
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars['DateTime'];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars['Boolean'];
};


/** Represents a comment. */
export type CommentUserContentEditsArgs = {
    after?: Maybe<Scalars['String']>;
    before?: Maybe<Scalars['String']>;
    first?: Maybe<Scalars['Int']>;
    last?: Maybe<Scalars['Int']>;
};

/** A comment author association with repository. */
export type CommentAuthorAssociation =
/** Author has been invited to collaborate on the repository. */
    | 'COLLABORATOR'
    /** Author has previously committed to the repository. */
    | 'CONTRIBUTOR'
    /** Author has not previously committed to GitHub. */
    | 'FIRST_TIMER'
    /** Author has not previously committed to the repository. */
    | 'FIRST_TIME_CONTRIBUTOR'
    /** Author is a placeholder for an unclaimed user. */
    | 'MANNEQUIN'
    /** Author is a member of the organization that owns the repository. */
    | 'MEMBER'
    /** Author has no association with the repository. */
    | 'NONE'
    /** Author is the owner of the repository. */
    | 'OWNER';

/** The possible errors that will prevent a user from updating a comment. */
export type CommentCannotUpdateReason =
/** Unable to create comment because repository is archived. */
    | 'ARCHIVED'
    /** You cannot update this comment */
    | 'DENIED'
    /** You must be the author or have write access to this repository to update this comment. */
    | 'INSUFFICIENT_ACCESS'
    /** Unable to create comment because issue is locked. */
    | 'LOCKED'
    /** You must be logged in to update this comment. */
    | 'LOGIN_REQUIRED'
    /** Repository is under maintenance. */
    | 'MAINTENANCE'
    /** At least one email address must be verified to update this comment. */
    | 'VERIFIED_EMAIL_REQUIRED';
// ...
export type Deletable = {
    /** Check if the current viewer can delete this object. */
    viewerCanDelete: Scalars['Boolean'];
};
// ...
export type IssueComment = Comment & Deletable & Minimizable & Node & Reactable & RepositoryNode & Updatable & UpdatableComment & {
    __typename?: 'IssueComment';
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** The body as Markdown. */
    body: Scalars['String'];
    /** The body rendered to HTML. */
    bodyHTML: Scalars['HTML'];
    /** The body rendered to text. */
    bodyText: Scalars['String'];
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars['DateTime'];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars['Boolean'];
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars['Int']>;
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    id: Scalars['ID'];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars['Boolean'];
    /** Returns whether or not a comment has been minimized. */
    isMinimized: Scalars['Boolean'];
    /** Identifies the issue associated with the comment. */
    issue: Issue;
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars['DateTime']>;
    /** Returns why the comment was minimized. */
    minimizedReason?: Maybe<Scalars['String']>;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars['DateTime']>;
    /**
     * Returns the pull request associated with the comment, if this comment was made on a
     * pull request.
     */
    pullRequest?: Maybe<PullRequest>;
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** The repository associated with this node. */
    repository: Repository;
    /** The HTTP path for this issue comment */
    resourcePath: Scalars['URI'];
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars['DateTime'];
    /** The HTTP URL for this issue comment */
    url: Scalars['URI'];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Check if the current viewer can delete this object. */
    viewerCanDelete: Scalars['Boolean'];
    /** Check if the current viewer can minimize this object. */
    viewerCanMinimize: Scalars['Boolean'];
    /** Can user react to this subject */
    viewerCanReact: Scalars['Boolean'];
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars['Boolean'];
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars['Boolean'];
};
// ...
export type IssueCommentEdge = {
    __typename?: 'IssueCommentEdge';
    /** A cursor for use in pagination. */
    cursor: Scalars['String'];
    /** The item at the end of the edge. */
    node?: Maybe<IssueComment>;
};
// ...
export type Minimizable = {
    /** Returns whether or not a comment has been minimized. */
    isMinimized: Scalars['Boolean'];
    /** Returns why the comment was minimized. */
    minimizedReason?: Maybe<Scalars['String']>;
    /** Check if the current viewer can minimize this object. */
    viewerCanMinimize: Scalars['Boolean'];
};
// ...
export type Node = {
    /** ID of the object. */
    id: Scalars['ID'];
};
// ...
export type Reactable = {
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars['Int']>;
    id: Scalars['ID'];
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** Can user react to this subject */
    viewerCanReact: Scalars['Boolean'];
};


/** Represents a subject that can be reacted on. */
export type ReactableReactionsArgs = {
    after?: Maybe<Scalars['String']>;
    before?: Maybe<Scalars['String']>;
    content?: Maybe<ReactionContent>;
    first?: Maybe<Scalars['Int']>;
    last?: Maybe<Scalars['Int']>;
    orderBy?: Maybe<ReactionOrder>;
};

/** The connection type for User. */
export type ReactingUserConnection = {
    __typename?: 'ReactingUserConnection';
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<ReactingUserEdge>>>;
    /** A list of nodes. */
    nodes?: Maybe<Array<Maybe<User>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /** Identifies the total count of items in the connection. */
    totalCount: Scalars['Int'];
};
// ...
export type ReactionConnection = {
    __typename?: 'ReactionConnection';
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<ReactionEdge>>>;
    /** A list of nodes. */
    nodes?: Maybe<Array<Maybe<Reaction>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /** Identifies the total count of items in the connection. */
    totalCount: Scalars['Int'];
    /** Whether or not the authenticated user has left a reaction on the subject. */
    viewerHasReacted: Scalars['Boolean'];
};

/** Emojis that can be attached to Issues, Pull Requests and Comments. */
export type ReactionContent =
/** Represents the `:confused:` emoji. */
    | 'CONFUSED'
    /** Represents the `:eyes:` emoji. */
    | 'EYES'
    /** Represents the `:heart:` emoji. */
    | 'HEART'
    /** Represents the `:hooray:` emoji. */
    | 'HOORAY'
    /** Represents the `:laugh:` emoji. */
    | 'LAUGH'
    /** Represents the `:rocket:` emoji. */
    | 'ROCKET'
    /** Represents the `:-1:` emoji. */
    | 'THUMBS_DOWN'
    /** Represents the `:+1:` emoji. */
    | 'THUMBS_UP';

/** An edge in a connection. */
export type ReactionEdge = {
    __typename?: 'ReactionEdge';
    /** A cursor for use in pagination. */
    cursor: Scalars['String'];
    /** The item at the end of the edge. */
    node?: Maybe<Reaction>;
};

/** A group of emoji reactions to a particular piece of content. */
export type ReactionGroup = {
    __typename?: 'ReactionGroup';
    /** Identifies the emoji reaction. */
    content: ReactionContent;
    /** Identifies when the reaction was created. */
    createdAt?: Maybe<Scalars['DateTime']>;
    /** Reactors to the reaction subject with the emotion represented by this reaction group. */
    reactors: ReactorConnection;
    /** The subject that was reacted to. */
    subject: Reactable;
    /**
     * Users who have reacted to the reaction subject with the emotion represented by this reaction group
     * @deprecated Reactors can now be mannequins, bots, and organizations. Use the `reactors` field instead. Removal on 2021-10-01 UTC.
     */
    users: ReactingUserConnection;
    /** Whether or not the authenticated user has left a reaction on the subject. */
    viewerHasReacted: Scalars['Boolean'];
};
// ...
export type ReactorConnection = {
    __typename?: 'ReactorConnection';
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<ReactorEdge>>>;
    /** A list of nodes. */
    nodes?: Maybe<Array<Maybe<Reactor>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /** Identifies the total count of items in the connection. */
    totalCount: Scalars['Int'];
};
// ...
export type RepositoryNode = {
    /** The repository associated with this node. */
    repository: Repository;
};
// ...
export type Updatable = {
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars['Boolean'];
};

/** Comments that can be updated. */
export type UpdatableComment = {
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
};
// ...
export type UserContentEditConnection = {
    __typename?: 'UserContentEditConnection';
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<UserContentEditEdge>>>;
    /** A list of nodes. */
    nodes?: Maybe<Array<Maybe<UserContentEdit>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /** Identifies the total count of items in the connection. */
    totalCount: Scalars['Int'];
};

/** An edge in a connection. */
export type UserContentEditEdge = {
    __typename?: 'UserContentEditEdge';
    /** A cursor for use in pagination. */
    cursor: Scalars['String'];
    /** The item at the end of the edge. */
    node?: Maybe<UserContentEdit>;
};